# Backend/api/openapi/teams-notification-api.yaml
openapi: 3.0.3
info:
  title: Teams Notification API (Go implementation)
  version: 1.0.0
  description: |
    This specification documents the REST endpoints implemented in the Teams Notification Go service.
    Includes all available routes for internal and external APIs.

    ## API Structure
    - **External APIs** (`/api/v1`): Public APIs for external clients
    - **Internal APIs** (`/internal/v1`): Internal management APIs
    - **System APIs**: Health, metrics, and configuration endpoints

    ## Recent Updates (v1.0.0)
    - **Unified Configuration**: Environment-based configuration system
    - **Error Handling**: Standardized error responses with proper HTTP status codes
    - **Token Management**: Redis-backed token caching for improved performance
    - **Monitoring**: Enhanced health checks and system metrics
    - **Logging**: Structured logging with configurable levels and formats
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.teamsnotify.com
    description: Production server
paths:
...
  /config:
    get:
      summary: Get system configuration
      tags: [System]
      description: Returns current system configuration (non-sensitive values only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: System configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /config/validate:
    post:
      summary: Validate system configuration
      tags: [System]
      description: Validates the current system configuration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Configuration validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigValidationResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  # External API endpoints (/api/v1)
...
  /internal/v1/projects:
    get:
      summary: List projects
      tags: [Internal API - Projects]
      description: Get a list of projects with pagination
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/SearchParam"
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"

    post:
      summary: Create project
      tags: [Internal API - Projects]
      description: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
      responses:
        "201":
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
  /internal/v1/projects/{id}:
    get:
      summary: Get project
      tags: [Internal API - Projects]
      description: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      summary: Update project
      tags: [Internal API - Projects]
      description: Update project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdateRequest"
      responses:
        "200":
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      summary: Delete project
      tags: [Internal API - Projects]
      description: Delete project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Project deleted successfully
        "404":
          $ref: "#/components/responses/NotFoundError"
  /internal/v1/projects/{id}/limits:
    patch:
      summary: Update project limits
      tags: [Internal API - Projects]
      description: Update project limits by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectLimitsUpdateRequest"
      responses:
        "200":
          description: Project limits updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /internal/v1/projects/company/{companyId}:
    get:
      summary: Get projects by company
      tags: [Internal API - Projects]
      description: Get projects by company ID
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of projects in company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /internal/v1/projects/key/{keyName}:
    get:
      summary: Get project by key name
      tags: [Internal API - Projects]
      description: Get project by notify key name
      parameters:
        - name: keyName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # Notifications endpoints
...
  /internal/v1/notifications/project/{projectId}:
    get:
      summary: Get notifications by project
      tags: [Internal API - Notifications]
      description: Get notifications by project ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of notifications for project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /internal/v1/destinations/project/{projectId}:
    get:
      summary: Get destinations by project
      tags: [Internal API - Destinations]
      description: Get destinations by project ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of destinations for project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationListResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /internal/v1/billing/usage/project/{projectId}:
    get:
      summary: Get project usage
      tags: [Internal API - Billing]
      description: Get usage records for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project usage records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageRecordsResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /internal/v1/billing/project/{projectId}:
    get:
      summary: Get project billing
      tags: [Internal API - Billing]
      description: Get project billing information
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project billing information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectBillingResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      summary: Update project billing
      tags: [Internal API - Billing]
      description: Update project billing information
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectBillingUpdateRequest"
      responses:
        "200":
          description: Project billing updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectBillingResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Number of items to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SearchParam:
      name: search
      in: query
      description: Search term
      required: false
      schema:
        type: string
...
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error information
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
    ProjectBillingUpdateRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [credit_card, bank_transfer, paypal]
        billingCycle:
          type: string
          enum: [monthly, yearly]
        nextBillingDate:
          type: string
          format: date
        billingStatus:
          type: string
          enum: [active, suspended, cancelled]
    PaginationResponse:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        pages:
          type: integer

    # Bot Framework schemas
...
    DestinationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [personal, group, channel]
        conversationId:
          type: string
        displayName:
          type: string
        tenantId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    # Configuration schemas
    ConfigResponse:
      type: object
      properties:
        environment:
          type: string
        version:
          type: string
        features:
          type: object
          additionalProperties: true
    ConfigValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string

    # User schemas
...
    ProjectCreateRequest:
      type: object
      required:
        - companyId
        - notifyKey
        - description
        - createdBy
      properties:
        companyId:
          type: string
          format: uuid
          description: Company ID
        notifyKey:
          type: string
          description: Project notify key
        description:
          type: string
          description: Project description
        dailyLimit:
          type: integer
          minimum: 0
          description: Daily notification limit
        monthlyLimit:
          type: integer
          minimum: 0
          description: Monthly notification limit
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
          description: Project priority
        createdBy:
          type: string
          format: uuid
          description: Creator user ID
    ProjectUpdateRequest:
      type: object
      properties:
        notifyKey:
          type: string
        description:
          type: string
        dailyLimit:
          type: integer
          minimum: 0
        monthlyLimit:
          type: integer
          minimum: 0
        priority:
          type: string
          enum: [low, normal, high]
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        notifyKey:
          type: string
        description:
          type: string
        dailyLimit:
          type: integer
        monthlyLimit:
          type: integer
        priority:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProjectResponse"
        pagination:
          $ref: "#/components/schemas/PaginationResponse"
    ProjectLimitsUpdateRequest:
      type: object
      properties:
        dailyLimit:
          type: integer
          minimum: 0
        monthlyLimit:
          type: integer
          minimum: 0

    # Notification schemas
...
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        messageType:
          type: string
        content:
          type: string
        mentions:
          type: array
          items:
            type: string
        priority:
          type: string
        status:
          type: string
        errorMessage:
          type: string
        destinationsSent:
          type: integer
        destinationsFailed:
          type: integer
        totalDestinations:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sentAt:
          type: string
          format: date-time
    NotificationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NotificationResponse"
        pagination:
          $ref: "#/components/schemas/PaginationResponse"
...
    DestinationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DestinationResponse"
        pagination:
          $ref: "#/components/schemas/PaginationResponse"

    # Bot schemas
...
    UsageRecordResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        notificationCount:
          type: integer
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    UsageRecordsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UsageRecordResponse"
        pagination:
          $ref: "#/components/schemas/PaginationResponse"
...
    ProjectBillingResponse:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        planName:
          type: string
        currentUsage:
          type: integer
        monthlyLimit:
          type: integer
        billingEnabled:
          type: boolean
        nextBillingDate:
          type: string
          format: date
...